+++ src/domain/community/contributor/user/FormikUserSelector/userSelectorQueries.graphql
